+++ CHANGELOG.md
@@ -1,3 +1,10 @@
+# v3.8.1 - 23.06.2026 (d.m.y)
+## Changed
+- Updated wording for extended click speed limit description.
+- Updated wording on task switcher -> Alt+Tab
+## Fixed
+- Fixed mouse cursor behaving erratically while application is running.
+
 # v3.8.0 - 21.06.2026 (d.m.y)
 ## New
 - SUPPORT FOR CUSTOM BACKGROUND IMAGES! (Settings → Appearance → Background Image).